refactor: move client related routes to 'routes/client.go'
