+++ workers/handlers/sync.go
@@ -1,4 +1,4 @@
-package client
+package handlers
 
 import "github.com/labstack/echo/v4"
