+++ workers/routes/proxy.go
@@ -1,4 +1,4 @@
-package proxy
+package routes
 
 import (
 	"github.com/labstack/echo/v4"
@@ -7,7 +7,10 @@ import (
 	"github.com/onsonr/sonr/nebula/components/home"
 )
 
-func RegisterViews(e *echo.Echo) {
+func RegisterProxyAPI(e *echo.Echo) {
+}
+
+func RegisterProxyViews(e *echo.Echo) {
 	e.GET("/", home.Route)
 	e.GET("/login", auth.LoginRoute)
 	e.GET("/register", auth.RegisterRoute)
